chore: update component name
